reorganize login screen
