--- src/screens/Auth/LogInScreen.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import AuthForm from './components/AuthForm';
-
-const LogInScreen = () => (
-    <AuthForm
-        authType="log-in"
-    />
-);
-
-export default LogInScreen;
