+++ src/screens/Register/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import AuthForm from './components/AuthForm';
+import AuthForm from '../../components/AuthForm';
 
 const RegisterScreen = ({ navigation }: any) => (
     <AuthForm
